Code review: moving utility method to static utility class
